refactor: FeedHeader에서 사용하는 옵저버 위치 변경
