+++ src/components/containers/Question/FeedHeader/FeedHeader.jsx
@@ -1,4 +1,4 @@
-import { forwardRef } from "react";
+import { useEffect, useRef, useState } from "react";
 import { Link } from "react-router-dom";
 
 import LogoImg from "@/assets/img/LogoImg";
@@ -10,11 +10,26 @@ import ScrollShareButtons from "@/components/containers/Question/FeedHeader/Scro
 
 import * as S from "@/components/containers/Question/FeedHeader/FeedHeader.style";
 
-const FeedHeader = forwardRef(({ subjectData, $isScroll }, ref) => {
+const FeedHeader = ({ subjectData }) => {
+  const [isHeaderVisible, setIsHeaderVisible] = useState(false);
+  const headerRef = useRef(null);
+
+  /* header scroll observer */
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        setIsHeaderVisible(!entry.isIntersecting);
+      },
+      { threshold: 0 },
+    );
+    if (headerRef.current) observer.observe(headerRef.current);
+    return () => observer.disconnect();
+  }, []);
+
   return (
     <>
       {/* 스크롤 안했을 때 헤더  */}
-      <S.MainHeader $hidden={$isScroll} ref={ref}>
+      <S.MainHeader $hidden={isHeaderVisible} ref={headerRef}>
         <S.ProfileContainer>
           <LogoImg width={170} as={Link} to="/" />
           <FeedProfile subjectData={subjectData} />
@@ -23,17 +38,17 @@ const FeedHeader = forwardRef(({ subjectData, $isScroll }, ref) => {
       </S.MainHeader>
 
       {/* 스크롤 헤더  */}
-      <S.ScrollContainer $visible={$isScroll}>
+      <S.ScrollContainer $visible={isHeaderVisible}>
         <S.PrevButton as={Link} to="/list">
           <ArrowLeftIcon size={44} />
         </S.PrevButton>
         <S.ScrollFeedProfile>
-          <FeedProfile subjectData={subjectData} $isScroll={$isScroll} />
+          <FeedProfile subjectData={subjectData} $isScroll={isHeaderVisible} />
         </S.ScrollFeedProfile>
-        <ScrollShareButtons $isScroll={$isScroll} />
+        <ScrollShareButtons $isScroll={isHeaderVisible} />
       </S.ScrollContainer>
     </>
   );
-});
+};
 
 export default FeedHeader;
